Free run - 5 - refactoring, 6

+++ src/freeRun/lc_05_29_Divide_Two_Integers/Solution.java
@@ -1,19 +1,19 @@
 package freeRun.lc_05_29_Divide_Two_Integers;
 
 public class Solution {
-    public int divide(int dividend, long divisor) {
+    public int divide(int dividend, int divisor) {
         int c = dividend < 0 && divisor > 0 || dividend >= 0 && divisor < 0 ? -1 : 1;
         long dd = Math.abs((long) dividend);
-        long dr = Math.abs((long)divisor);
-        if (divisor == 1){
+        long dr = Math.abs(divisor);
+        if (divisor == 1) {
             return dividend;
         }
         long res = 0;
         for (int i = 30; i >= 0; i--) {
-            long temp =  dr << i ;
+            long temp = dr << i;
             if (temp <= dd) {
-                res += (1<<i);
-                dd-=temp;
+                res += (1 << i);
+                dd -= temp;
             }
         }
         return (int) (c < 0 ? -res : res);
